separo HompePage componentes
