+++ src/pages/HomePage/HomePage.jsx
@@ -1,6 +1,62 @@
-import Slider from "../../components/Home/HomeSlider/Slider.jsx";
-const HomePage = () => {
-  return <Slider />;
-};
+import { useDispatch, useSelector } from "react-redux";
+import {Grid,Box} from "@mui/material";
+import { useEffect,useState } from "react";
+import { AccountApi } from "../../api/accountApi";
+import { addAccountArs, changeBalanceArs } from "../../redux/accountArsSlice";
+import { addAccountUsd } from "../../redux/accountUsdSlice";
+import UserInfoCard from "../../components/Home/UserInfoCard/UserInfoCard";
+import ActivitiesCard from "../../components/Home/ActivityCard/ActivitiesCard";
+import TransactionList from "../../components/Home/TransactionList/TransactionList";
 
-export default HomePage;
+
+
+export default function HomePage() {
+  const [currency, setCurrency] = useState("ARS");
+  const accountARS = useSelector((state) => state.accountArs);
+  const accountUSD = useSelector((state) => state.accountUsd);
+  const user = useSelector((state) => state.user);
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    AccountApi.accountInfo(user.id).then((response) => {
+      console.log(response.data);
+      response.data.map((account) => {
+        if (account.currency === "ARS") {
+          dispatch(addAccountArs(account));
+        } else {
+          dispatch(addAccountUsd(account));
+        }
+      });
+    });
+    AccountApi.balance().then((response) => {
+      dispatch(changeBalanceArs(response.data));
+    });
+  }, [dispatch, user]);
+
+  const handleForward = () => {
+    setCurrency(currency === "ARS" ? "USD" : "ARS");
+  };
+
+  return (
+    <>
+      <Box sx={{ flexGrow: 2, paddingLeft: "%" }}>
+        <Grid container spacing={2}>
+          <Grid item xs={1} />
+          <Grid item xs={1} />
+
+          <Grid item xs={2}>
+            <UserInfoCard
+              currency={currency}
+              accountARS={accountARS}
+              accountUSD={accountUSD}
+              user={user}
+              handleForward={handleForward}
+            />
+            <ActivitiesCard />
+          </Grid>
+          <TransactionList />
+        </Grid>
+      </Box>
+    </>
+  );
+}
